requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.0
    # via django
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via pytest
backcall==0.2.0
    # via ipython
backports-zoneinfo==0.2.1
    # via django
black==22.1.0
    # via ipython
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.0.3
    # via black
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
dj-database-url==0.5.0
    # via -r requirements.in
django==4.0.2
    # via
    #   -r requirements.in
    #   djangorestframework
    #   drf-yasg
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg==1.20.0
    # via -r requirements.in
executing==0.8.2
    # via stack-data
factory-boy==3.2.1
    # via
    #   -r requirements.in
    #   pytest-factoryboy
faker==12.3.0
    # via
    #   -r requirements.in
    #   factory-boy
idna==3.3
    # via requests
inflection==0.5.1
    # via
    #   drf-yasg
    #   pytest-factoryboy
iniconfig==1.1.1
    # via pytest
ipdb==0.13.9
    # via -r requirements.in
ipython==8.0.1
    # via ipdb
itypes==1.2.0
    # via coreapi
jedi==0.18.1
    # via ipython
jinja2==3.0.3
    # via coreschema
markupsafe==2.0.1
    # via jinja2
matplotlib-inline==0.1.3
    # via ipython
mypy-extensions==0.4.3
    # via black
numpy==1.22.2
    # via pandas
packaging==21.3
    # via
    #   drf-yasg
    #   pytest
pandas==1.4.1
    # via -r requirements.in
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==2.5.0
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.27
    # via ipython
psycopg2-binary==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via pytest
pygments==2.11.2
    # via ipython
pyparsing==3.0.7
    # via packaging
pytest==7.0.1
    # via
    #   pytest-django
    #   pytest-factoryboy
    #   pytest-mock
pytest-django==4.5.2
    # via -r requirements.in
pytest-factoryboy==2.1.0
    # via -r requirements.in
pytest-mock==3.7.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   faker
    #   pandas
python-decouple==3.6.0
    # via -r requirements.in
pytz==2021.3
    # via
    #   djangorestframework
    #   pandas
requests==2.27.1
    # via coreapi
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via django
stack-data==0.1.4
    # via ipython
toml==0.10.2
    # via ipdb
tomli==2.0.1
    # via
    #   black
    #   pytest
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.1.0
    # via black
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.8
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
